move trace storage、move browser
