+++ cmd/monitor/streaming/main.go
@@ -23,7 +23,7 @@ import (
 	_ "github.com/erda-project/erda/modules/monitor/alert/storage/alert-record"
 	_ "github.com/erda-project/erda/modules/monitor/apm/browser"
 	_ "github.com/erda-project/erda/modules/monitor/core/logs/storage"
-	_ "github.com/erda-project/erda/modules/monitor/trace/storage"
+	_ "github.com/erda-project/erda/modules/msp/apm/trace/storage"
 
 	// providers
 	_ "github.com/erda-project/erda-infra/providers/cassandra"
