+++ modules/msp/apm/browser/consumer.go
@@ -25,8 +25,8 @@ import (
 	"github.com/varstr/uaparser"
 
 	metrics "github.com/erda-project/erda/modules/core/monitor/metric"
-	"github.com/erda-project/erda/modules/monitor/apm/browser/timing"
 	"github.com/erda-project/erda/modules/monitor/utils"
+	"github.com/erda-project/erda/modules/msp/apm/browser/timing"
 )
 
 var (
