+++ modules/msp/apm/browser/timing/navigation_timing.go
