+++ modules/msp/apm/browser/timing/performance_timing.go
