+++ modules/msp/apm/trace/storage/consumer_test.go
@@ -22,7 +22,6 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	metrics "github.com/erda-project/erda/modules/core/monitor/metric"
-	"github.com/erda-project/erda/modules/monitor/trace"
 )
 
 // TestMetricToSpan .
@@ -147,7 +146,7 @@ func TestToInt64(t *testing.T) {
 
 // TestToSpan .
 func TestToSpan(t *testing.T) {
-	span := trace.Span{
+	span := Span{
 		TraceID:       "test-trace-id",
 		StartTime:     time.Now().Add(-1).UnixNano(),
 		SpanID:        "test-span-id",
