+++ modules/msp/apm/trace/storage/trace.go
@@ -11,7 +11,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package trace
+package storage
 
 // Span .
 type Span struct {
